+++ examples/misc/code-splitting/src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { curiProvider } from "@curi/react-dom";
 
-import Browser from "@hickory/browser";
+import { Browser } from "@hickory/browser";
 import { curi } from "@curi/router";
 
 import routes from "./routes";